+++ packages/react/src/index2.ts
@@ -1,0 +1,61 @@
+"use client";
+
+export {
+  Accordion,
+  AccordionItem,
+  type AccordionItemProps,
+  type AccordionProps,
+} from "@jamsr-ui/accordion";
+export { Alert, type AlertProps } from "@jamsr-ui/alert";
+export { AutoComplete, type AutoCompleteProps } from "@jamsr-ui/autocomplete";
+export { Avatar, type AvatarProps } from "@jamsr-ui/avatar";
+export { Badge, type BadgeProps } from "@jamsr-ui/badge";
+export {
+  Button,
+  type ButtonProps,
+  type UseButtonProps,
+} from "@jamsr-ui/button";
+export {
+  Card,
+  CardContent,
+  CardFooter,
+  CardHeader,
+  type CardContentProps,
+  type CardFooterProps,
+  type CardHeaderProps,
+  type CardProps,
+} from "@jamsr-ui/card";
+export { Checkbox, type CheckboxProps } from "@jamsr-ui/checkbox";
+export { Chip, type ChipProps } from "@jamsr-ui/chip";
+export {
+  Dialog,
+  DialogBody,
+  DialogContent,
+  DialogFooter,
+  DialogHeader,
+  type DialogBodyProps,
+  type DialogContentProps,
+  type DialogFooterProps,
+  type DialogHeaderProps,
+  type DialogProps,
+} from "@jamsr-ui/dialog";
+// export * from "@jamsr-ui/divider";
+// export * from "@jamsr-ui/header";
+// export * from "@jamsr-ui/image-upload";
+// export * from "@jamsr-ui/input";
+// export * from "@jamsr-ui/link";
+// export * from "@jamsr-ui/menu";
+// export * from "@jamsr-ui/otp-input";
+// export * from "@jamsr-ui/popover";
+// export * from "@jamsr-ui/progress";
+// export * from "@jamsr-ui/radio";
+// export * from "@jamsr-ui/ripple";
+// export * from "@jamsr-ui/select";
+// export * from "@jamsr-ui/skeleton";
+// export * from "@jamsr-ui/switch";
+// export * from "@jamsr-ui/tab";
+// export * from "@jamsr-ui/tooltip";
+// export * from "@jamsr-ui/typography";
+
+//
+export { JamsrUIProvider } from "@jamsr-ui/core";