+++ packages/react/tsup.config.ts
@@ -2,13 +2,14 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   entry: ["src"],
-  minify: true,
   format: ["cjs", "esm"],
+  dts: true,
   clean: true,
+  minify: true,
+  splitting: true,
   esbuildOptions(options) {
     options.banner = {
       js: '"use client"',
     };
   },
-  dts: true,
 });