+++ packages/components/accordion/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @jamsr-ui/accordion
 
+## 0.0.22
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/shared-icons@0.0.20
+  - @jamsr-ui/utils@0.0.19
+
+## 0.0.21
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/shared-icons@0.0.19
+  - @jamsr-ui/utils@0.0.18
+
 ## 0.0.20
 
 ### Patch Changes