+++ packages/components/card/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @jamsr-ui/card
 
+## 0.0.21
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/typography@0.0.21
+  - @jamsr-ui/progress@0.0.21
+  - @jamsr-ui/utils@0.0.19
+
+## 0.0.20
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/progress@0.0.20
+  - @jamsr-ui/typography@0.0.20
+  - @jamsr-ui/utils@0.0.18
+
 ## 0.0.19
 
 ### Patch Changes