+++ packages/components/dialog/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @jamsr-ui/dialog
 
+## 0.0.22
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/button@0.0.21
+  - @jamsr-ui/shared-icons@0.0.20
+  - @jamsr-ui/hooks@0.0.14
+  - @jamsr-ui/utils@0.0.19
+
+## 0.0.21
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/button@0.0.20
+  - @jamsr-ui/hooks@0.0.13
+  - @jamsr-ui/shared-icons@0.0.19
+  - @jamsr-ui/utils@0.0.18
+
 ## 0.0.20
 
 ### Patch Changes