+++ packages/components/header/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @jamsr-ui/header
 
+## 0.0.21
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/divider@0.0.21
+  - @jamsr-ui/utils@0.0.19
+
+## 0.0.20
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/divider@0.0.20
+  - @jamsr-ui/utils@0.0.18
+
 ## 0.0.19
 
 ### Patch Changes