+++ packages/components/image-upload/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @jamsr-ui/image-upload
 
+## 0.0.21
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.19
+
+## 0.0.20
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.18
+
 ## 0.0.19
 
 ### Patch Changes