+++ packages/config/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jamsr-ui/eslint-config
 
+## 0.0.13
+
+### Patch Changes
+
+- fix
+
+## 0.0.12
+
+### Patch Changes
+
+- fix
+
 ## 0.0.11
 
 ### Patch Changes