+++ packages/config/typescript-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jamsr-ui/typescript-config
 
+## 0.0.14
+
+### Patch Changes
+
+- fix
+
+## 0.0.13
+
+### Patch Changes
+
+- fix
+
 ## 0.0.12
 
 ### Patch Changes