+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jamsr-ui/core
 
+## 0.0.4
+
+### Patch Changes
+
+- fix
+
+## 0.0.3
+
+### Patch Changes
+
+- fix
+
 ## 0.0.2
 
 ### Patch Changes