+++ packages/react/CHANGELOG.md
@@ -1,5 +1,77 @@
 # @jamsr-ui/react
 
+## 0.0.28
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/autocomplete@0.0.21
+  - @jamsr-ui/image-upload@0.0.21
+  - @jamsr-ui/typography@0.0.21
+  - @jamsr-ui/accordion@0.0.22
+  - @jamsr-ui/otp-input@0.0.21
+  - @jamsr-ui/checkbox@0.0.21
+  - @jamsr-ui/progress@0.0.21
+  - @jamsr-ui/skeleton@0.0.21
+  - @jamsr-ui/divider@0.0.21
+  - @jamsr-ui/popover@0.0.22
+  - @jamsr-ui/tooltip@0.0.21
+  - @jamsr-ui/avatar@0.0.21
+  - @jamsr-ui/button@0.0.21
+  - @jamsr-ui/dialog@0.0.22
+  - @jamsr-ui/header@0.0.21
+  - @jamsr-ui/ripple@0.0.21
+  - @jamsr-ui/select@0.0.22
+  - @jamsr-ui/switch@0.0.21
+  - @jamsr-ui/alert@0.0.21
+  - @jamsr-ui/badge@0.0.21
+  - @jamsr-ui/input@0.0.21
+  - @jamsr-ui/radio@0.0.21
+  - @jamsr-ui/card@0.0.21
+  - @jamsr-ui/chip@0.0.21
+  - @jamsr-ui/link@0.0.22
+  - @jamsr-ui/menu@0.0.22
+  - @jamsr-ui/tab@0.0.21
+  - @jamsr-ui/utils@0.0.19
+  - @jamsr-ui/core@0.0.4
+
+## 0.0.27
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/accordion@0.0.21
+  - @jamsr-ui/alert@0.0.20
+  - @jamsr-ui/autocomplete@0.0.20
+  - @jamsr-ui/avatar@0.0.20
+  - @jamsr-ui/badge@0.0.20
+  - @jamsr-ui/button@0.0.20
+  - @jamsr-ui/card@0.0.20
+  - @jamsr-ui/checkbox@0.0.20
+  - @jamsr-ui/chip@0.0.20
+  - @jamsr-ui/dialog@0.0.21
+  - @jamsr-ui/divider@0.0.20
+  - @jamsr-ui/header@0.0.20
+  - @jamsr-ui/image-upload@0.0.20
+  - @jamsr-ui/input@0.0.20
+  - @jamsr-ui/link@0.0.21
+  - @jamsr-ui/menu@0.0.21
+  - @jamsr-ui/otp-input@0.0.20
+  - @jamsr-ui/popover@0.0.21
+  - @jamsr-ui/progress@0.0.20
+  - @jamsr-ui/radio@0.0.20
+  - @jamsr-ui/ripple@0.0.20
+  - @jamsr-ui/select@0.0.21
+  - @jamsr-ui/skeleton@0.0.20
+  - @jamsr-ui/switch@0.0.20
+  - @jamsr-ui/tab@0.0.20
+  - @jamsr-ui/tooltip@0.0.20
+  - @jamsr-ui/typography@0.0.20
+  - @jamsr-ui/core@0.0.3
+  - @jamsr-ui/utils@0.0.18
+
 ## 0.0.26
 
 ### Patch Changes