+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jamsr-ui/theme
 
+## 0.0.19
+
+### Patch Changes
+
+- fix
+
+## 0.0.18
+
+### Patch Changes
+
+- fix
+
 ## 0.0.17
 
 ### Patch Changes